refactor: Use dart:js_interop for browser client (#35)
Create an extension type for `XMLHttpRequest` and use that instead of `HttpRequest` from `dart:html`. The `web` package is not used because it does not expose all of the methods that were being used in the original client. Also simplify things to not use a `Stream` and use `addEventListener` directly.